GrayScale as a separate dist library
