+++ examples/example_w_worker.html
@@ -7,6 +7,7 @@
   <meta http-equiv="X-UA-Compatible" content="chrome=1">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <script src="../dist/WebARKit.js"></script>
+  <script src="../dist/GrayScale.js"></script>
   <script type="text/javascript" src="stats.min.js"></script>
   <script type="module" src="index_w_worker.js"></script>
   <link rel="stylesheet" href="css/style.css">
