+++ src/index.js
@@ -1,7 +1,5 @@
 import WebARKitController from './WebARKitController'
-import {GrayScaleMedia} from './utils/Grayscale'
 
 export default {
-  GrayScaleMedia,
   WebARKitController
 }
