+++ src/utils/GrayScale.js
@@ -1,0 +1,130 @@
+function isMobile() {
+    let mobile = false;
+    (function(a){if(/(android|bb\d+|meego).+mobile|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(hone|od)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(ob|in)i|palm( os)?|phone|p(ixi|re)\/|plucker|pocket|psp|series(4|6)0|symbian|treo|up\.(browser|link)|vodafone|wap|windows ce|xda|xiino/i.test(a)||/1207|6310|6590|3gso|4thp|50[1-6]i|770s|802s|a wa|abac|ac(er|oo|s\-)|ai(ko|rn)|al(av|ca|co)|amoi|an(ex|ny|yw)|aptu|ar(ch|go)|as(te|us)|attw|au(di|\-m|r |s )|avan|be(ck|ll|nq)|bi(lb|rd)|bl(ac|az)|br(e|v)w|bumb|bw\-(n|u)|c55\/|capi|ccwa|cdm\-|cell|chtm|cldc|cmd\-|co(mp|nd)|craw|da(it|ll|ng)|dbte|dc\-s|devi|dica|dmob|do(c|p)o|ds(12|\-d)|el(49|ai)|em(l2|ul)|er(ic|k0)|esl8|ez([4-7]0|os|wa|ze)|fetc|fly(\-|_)|g1 u|g560|gene|gf\-5|g\-mo|go(\.w|od)|gr(ad|un)|haie|hcit|hd\-(m|p|t)|hei\-|hi(pt|ta)|hp( i|ip)|hs\-c|ht(c(\-| |_|a|g|p|s|t)|tp)|hu(aw|tc)|i\-(20|go|ma)|i230|iac( |\-|\/)|ibro|idea|ig01|ikom|im1k|inno|ipaq|iris|ja(t|v)a|jbro|jemu|jigs|kddi|keji|kgt( |\/)|klon|kpt |kwc\-|kyo(c|k)|le(no|xi)|lg( g|\/(k|l|u)|50|54|\-[a-w])|libw|lynx|m1\-w|m3ga|m50\/|ma(te|ui|xo)|mc(01|21|ca)|m\-cr|me(rc|ri)|mi(o8|oa|ts)|mmef|mo(01|02|bi|de|do|t(\-| |o|v)|zz)|mt(50|p1|v )|mwbp|mywa|n10[0-2]|n20[2-3]|n30(0|2)|n50(0|2|5)|n7(0(0|1)|10)|ne((c|m)\-|on|tf|wf|wg|wt)|nok(6|i)|nzph|o2im|op(ti|wv)|oran|owg1|p800|pan(a|d|t)|pdxg|pg(13|\-([1-8]|c))|phil|pire|pl(ay|uc)|pn\-2|po(ck|rt|se)|prox|psio|pt\-g|qa\-a|qc(07|12|21|32|60|\-[2-7]|i\-)|qtek|r380|r600|raks|rim9|ro(ve|zo)|s55\/|sa(ge|ma|mm|ms|ny|va)|sc(01|h\-|oo|p\-)|sdk\/|se(c(\-|0|1)|47|mc|nd|ri)|sgh\-|shar|sie(\-|m)|sk\-0|sl(45|id)|sm(al|ar|b3|it|t5)|so(ft|ny)|sp(01|h\-|v\-|v )|sy(01|mb)|t2(18|50)|t6(00|10|18)|ta(gt|lk)|tcl\-|tdg\-|tel(i|m)|tim\-|t\-mo|to(pl|sh)|ts(70|m\-|m3|m5)|tx\-9|up(\.b|g1|si)|utst|v400|v750|veri|vi(rg|te)|vk(40|5[0-3]|\-v)|vm40|voda|vulc|vx(52|53|60|61|70|80|81|83|85|98)|w3c(\-| )|webc|whit|wi(g |nc|nw)|wmlb|wonu|x700|yas\-|your|zeto|zte\-/i.test(a.substr(0,4))) mobile = true;})(navigator.userAgent||navigator.vendor||window.opera);
+    return mobile;
+};
+
+export class GrayScaleMedia
+{
+    constructor(source, width, height, canvas) {
+        this._source = source;
+        this._width = width;
+        this._height = height;
+        this._canvas = canvas ? canvas : document.createElement("canvas");
+        this._canvas.width = width;
+        this._canvas.height = height;
+
+        this._flipImageProg = require("./shaders/flip-image.glsl");
+        this._grayscaleProg = require("./shaders/grayscale.glsl");
+        this.glReady = false;
+        this.initGL(this._flipImageProg, this._grayscaleProg);
+    }
+
+    initGL(vertShaderSource, fragShaderSource) {
+        this.gl = this._canvas.getContext("webgl");
+
+        this.gl.viewport(0, 0, this.gl.drawingBufferWidth, this.gl.drawingBufferHeight);
+        this.gl.clearColor(0.1, 0.1, 0.1, 1.0);
+        this.gl.clear(this.gl.COLOR_BUFFER_BIT);
+
+        const vertShader = this.gl.createShader(this.gl.VERTEX_SHADER);
+        const fragShader = this.gl.createShader(this.gl.FRAGMENT_SHADER);
+
+        this.gl.shaderSource(vertShader, vertShaderSource);
+        this.gl.shaderSource(fragShader, fragShaderSource);
+
+        this.gl.compileShader(vertShader);
+        this.gl.compileShader(fragShader);
+
+        const program = this.gl.createProgram();
+        this.gl.attachShader(program, vertShader);
+        this.gl.attachShader(program, fragShader);
+
+        this.gl.linkProgram(program);
+
+        this.gl.useProgram(program);
+
+        const vertices = new Float32Array([
+            -1, -1,
+            -1,  1,
+             1,  1,
+            -1, -1,
+             1,  1,
+             1, -1,
+        ]);
+
+        const buffer = this.gl.createBuffer();
+        this.gl.bindBuffer(this.gl.ARRAY_BUFFER, buffer);
+        this.gl.bufferData(this.gl.ARRAY_BUFFER, vertices, this.gl.STATIC_DRAW);
+
+        const positionLocation = this.gl.getAttribLocation(program, "position");
+        this.gl.vertexAttribPointer(positionLocation, 2, this.gl.FLOAT, false, 0, 0);
+        this.gl.enableVertexAttribArray(positionLocation);
+
+        this.flipLocation = this.gl.getUniformLocation(program, "flipY");
+
+        const texture = this.gl.createTexture();
+        this.gl.activeTexture(this.gl.TEXTURE0);
+        this.gl.bindTexture(this.gl.TEXTURE_2D, texture);
+
+        // if either dimension of image is not a power of 2
+        this.gl.texParameteri(this.gl.TEXTURE_2D, this.gl.TEXTURE_WRAP_S, this.gl.CLAMP_TO_EDGE);
+        this.gl.texParameteri(this.gl.TEXTURE_2D, this.gl.TEXTURE_WRAP_T, this.gl.CLAMP_TO_EDGE);
+        this.gl.texParameteri(this.gl.TEXTURE_2D, this.gl.TEXTURE_MIN_FILTER, this.gl.LINEAR);
+        this.gl.texParameteri(this.gl.TEXTURE_2D, this.gl.TEXTURE_MAG_FILTER, this.gl.LINEAR);
+
+        this.glReady = true;
+        this.pixelBuf = new Uint8Array(this.gl.drawingBufferWidth * this.gl.drawingBufferHeight * 4);
+        this.grayBuf = new Uint8Array(this.gl.drawingBufferWidth * this.gl.drawingBufferHeight);
+    }
+
+    getFrame() {
+        if (!this.glReady) return undefined;
+
+        this.gl.uniform1f(this.flipLocation, -1); // flip image
+        this.gl.texImage2D(this.gl.TEXTURE_2D, 0, this.gl.RGBA, this.gl.RGBA, this.gl.UNSIGNED_BYTE, this._source);
+        this.gl.drawArrays(this.gl.TRIANGLES, 0, 6);
+
+        this.gl.readPixels(0, 0, this.gl.drawingBufferWidth, this.gl.drawingBufferHeight, this.gl.RGBA, this.gl.UNSIGNED_BYTE, this.pixelBuf);
+
+        let j = 0;
+        for (let i = 0; i < this.pixelBuf.length; i += 4) {
+            this.grayBuf[j] = this.pixelBuf[i];
+            j++;
+        }
+        return this.grayBuf;
+    }
+
+    requestStream() {
+        return new Promise((resolve, reject) => {
+            if (!navigator.mediaDevices || !navigator.mediaDevices.getUserMedia)
+                return reject();
+
+            // Hack for mobile browsers: aspect ratio is flipped.
+            var aspect = this._width / this._height;
+            if (isMobile()) {
+                aspect = 1 / aspect;
+            }
+
+            navigator.mediaDevices.getUserMedia({
+                audio: false,
+                video: {
+                    width: { ideal: this._width },
+                    height: { ideal: this._height },
+                    aspectRatio: { ideal: aspect },
+                    facingMode: "environment"
+                }
+            })
+            .then(stream => {
+                this._source.srcObject = stream;
+                this._source.onloadedmetadata = e => {
+                    this._source.play();
+                    resolve(this._source);
+                };
+            })
+            .catch(err => {
+                reject(err);
+            });
+        });
+    }
+}
